Update gui.py
In this commit is what I changed to get this to work on macOS tested on Apple Silicon. I REM'd out the "app = QApplication(sys.argv + ["-platform", "windows:darkmode=2"])" and added to use cocoa. There may be a better way to code this so that it works with Windows and macOS but I am not much of a developer. 

Also added code to allow the window to be resized and added a scroll bar when necessary. 

With these changes, I can start torf via cmdl.